+++ src/libs/info/windows/scan_port.cpp
@@ -26,7 +26,7 @@
   #include "network/scan_port.h"
 
   namespace bsscan {
-
+    // STATUS OF PORT
     std::string PORT::getStatus(int port) {
       std::cout << port << endl;
       if (port_is_open("127.0.0.1", port))
@@ -34,79 +34,126 @@
       else
         return "CLOSED";
     }
-
+    // SERVICE OF PORT
     std::string PORT::getService(int port) {
       return "TODO";
     }
-
+    // ALL PORTS
     std::vector<PORT> getAllPorts() {
       std::string address = "127.0.0.1";
-      std::string port_list = "1-48556";
       std::vector<int> new_ports;
       std::vector<PORT> ports;
-      new_ports = ports_list(port_list);
+      new_ports = ports_list("1-48556");
       std::cout << "\nScanning...\n " << endl;
       for (int port : new_ports) {
         ports.push_back(PORT());
-
-        std::cout << "Port " << port << " : ";
+        std::cout << "Port " << port << ":  ";
         if (port_is_open(address, port)) {
           std::cout << "OPEN\n";
           ports.back()._s_status = "OPEN";
           ports.back()._n_number = port;
-        
         }else {
           std::cout << "CLOSED\n";
           ports.back()._s_status = "CLOSED";
           ports.back()._n_number = port;
         }
       }
-
       return ports;
     };
-
+    // ONLY OPEN
     std::vector<PORT> getAllPortsOpen() {
       std::string address = "127.0.0.1";
-      std::string port_list = "1-48556";
       std::vector<int> new_ports;
       std::vector<PORT> ports;
-      new_ports = ports_list(port_list);
+      new_ports = ports_list("1-48556");
       std::cout << "\nScanning...\n " << endl;
       for (int port : new_ports) {
         ports.push_back(PORT());
-
         if (port_is_open(address, port)) {
-          std::cout << "Port " << port << " : ";
+          std::cout << "Port " << port << ":  ";
           std::cout << "OPEN\n";
           ports.back()._s_status = "OPEN";
           ports.back()._n_number = port;
         }
       }
-
       return ports;
     };
-
+    // ONLY CLOSED
     std::vector<PORT> getAllPortsClosed() {
       std::string address = "127.0.0.1";
-      std::string port_list = "1-48556";
       std::vector<int> new_ports;
       std::vector<PORT> ports;
-      new_ports = ports_list(port_list);
+      new_ports = ports_list("1-48556");
       std::cout << "\nScanning...\n " << endl;
       for (int port : new_ports) {
         ports.push_back(PORT());
-
         if (!port_is_open(address, port)) {
-          std::cout << "Port " << port << " : ";
+          std::cout << "Port " << port << ":  ";
           std::cout << "CLOSED\n";
           ports.back()._s_status = "CLOSED";
           ports.back()._n_number = port;
         }
       }
-
       return ports;
     };
-
+    // SEQUENTIAL PORTS
+    std::vector<PORT> getSeqPorts(string seq) {
+      std::string address = "127.0.0.1";
+      std::vector<int> new_ports;
+      std::vector<PORT> ports;
+      new_ports = ports_list(seq);
+      std::cout << "\nScanning...\n " << endl;
+      for (int port : new_ports) {
+        ports.push_back(PORT());
+        std::cout << "Port " << port << ":  ";
+        if (port_is_open(address, port)) {
+          std::cout << "OPEN\n";
+          ports.back()._s_status = "OPEN";
+          ports.back()._n_number = port;
+        } else {
+          std::cout << "CLOSED\n";
+          ports.back()._s_status = "CLOSED";
+          ports.back()._n_number = port;
+        }
+      }
+      return ports;
+    };
+    // SEQUENTIAL ONLY CLOSED
+    std::vector<PORT> getSeqPortsClosed(string seq) {
+      std::string address = "127.0.0.1";
+      std::vector<int> new_ports;
+      std::vector<PORT> ports;
+      new_ports = ports_list(seq);
+      std::cout << "\nScanning...\n " << endl;
+      for (int port : new_ports) {
+        ports.push_back(PORT());
+        if (!port_is_open(address, port)) {
+          std::cout << "Port " << port << ":  ";
+          std::cout << "CLOSED\n";
+          ports.back()._s_status = "CLOSED";
+          ports.back()._n_number = port;
+        }
+      }
+      return ports;
+    };
+    // SEQUENTIAL ONLY OPEN
+    std::vector<PORT> getSeqPortsOpen(string seq) {
+      std::string address = "127.0.0.1";
+      std::vector<int> new_ports;
+      std::vector<PORT> ports;
+      new_ports = ports_list(seq);
+      std::cout << "\nScanning...\n " << endl;
+      for (int port : new_ports) {
+        ports.push_back(PORT());
+        if (port_is_open(address, port)) {
+          std::cout << "Port " << port << ":  ";
+          std::cout << "OPEN\n";
+          ports.back()._s_status = "OPEN";
+          ports.back()._n_number = port;
+        }
+      }
+      return ports;
+    };
 
   };
 #endif  // END BSSCAN_WINDOWS